refactor(upload): extract ArtworkField, drop the loq relax
the artwork field moves whole into ArtworkField.svelte — markup, styles,
validation, and the single-setter slot logic — with `bind:file` as its one
seam to the page. the page shrinks back under its original loq budget, so
the loq.toml bump is reverted, and the page-scope button-rule overrides
(width/weight/hover undo) are deleted rather than carried: component scope
is out of the page rule's reach, which is the fix the overrides were
approximating.
