+++ loq.toml
@@ -175,7 +175,7 @@ max_lines = 1299
 
 [[rules]]
 path = "frontend/src/routes/upload/+page.svelte"
-max_lines = 1162
+max_lines = 983
 
 [[rules]]
 path = "services/moderation/src/admin.rs"
